+++ experiments/inf_sing.py
