+++ experiments/stats/sub1_mean.npy
