+++ experiments/stats/sub1_var.npy
